+++ packages/fyndiq-component-loader/progressbar.css
@@ -1,0 +1,13 @@
+@import "fyndiq-styles-colors/colors.css";
+
+.progressbar {
+  width: 100%;
+  height: 6px;
+  border-radius: 6px;
+}
+
+.progress {
+  height: 100%;
+  min-width: 6px;
+  border-radius: 6px;
+}
